fix: 修复 token 类型不一致问题
- 将 model/aimodel.go 中 Usage 结构体的 token 字段类型从 int 改为 int64
- 移除 service/llm.go 中未使用的 context 包导入
- 统一所有 token 相关变量的类型为 int64

+++ backend/internal/model/aimodel.go
@@ -126,9 +126,9 @@ type LLMResponse struct {
 		FinishReason string `json:"finish_reason"`
 	} `json:"choices"`
 	Usage struct {
-		PromptTokens     int `json:"prompt_tokens"`
-		CompletionTokens int `json:"completion_tokens"`
-		TotalTokens      int `json:"total_tokens"`
+		PromptTokens     int64 `json:"prompt_tokens"`
+		CompletionTokens int64 `json:"completion_tokens"`
+		TotalTokens      int64 `json:"total_tokens"`
 	} `json:"usage"`
 }
